+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 2023.7.5 (2023-08-01)
+
+- Add SystemInformation to client api
+- Send credentials on XmlRPC api only when authentication is enabled in CCU
+
 # Version 2023.7.4 (2023-07-18)
 
 - Update requirements